Update MethodCallHandlerImpl.java
number casted to int

+++ android/src/main/java/yeniellandestoy/native_shared_preferences/MethodCallHandlerImpl.java
@@ -93,7 +93,7 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
                         key, BIG_INTEGER_PREFIX + integerValue.toString(Character.MAX_RADIX)),
                 result);
           } else {
-            commitAsync(preferences.edit().putInt(key, number), result);
+            commitAsync(preferences.edit().putInt(key, number.intValue()), result);
           }
           break;
         case "setString":
